musicgen_pywebview_requirements.txt
# Dependencies required to package the MusicGen PyWebview executable
pywebview>=4.4
gradio>=3.50
